changed mass to double

+++ planetary-playground-services/src/main/java/kerbonautas/planetaryplaygroundservices/Planeta.java
@@ -6,12 +6,12 @@ public class Planeta {
 	private double temperature;
 	private int numberOfSatellites;
 	private boolean isSolid;
-	private int mass;
+	private double mass;
 	private boolean haveAtmosphere;
 	private String name;
 	private double pressure;
 	
-	public Planeta(int rad, double temp, int sat, boolean solid, int mass, boolean haveAtmosphere, String name, double pressure) {
+	public Planeta(int rad, double temp, int sat, boolean solid, double mass, boolean haveAtmosphere, String name, double pressure) {
 		setRadious(rad);
 		setTemperature(temp);
 		setNumberOfSatellites(sat);
@@ -26,7 +26,7 @@ public class Planeta {
 	public void setTemperature(double i) {temperature = i;}
 	public void setNumberOfSatellites(int i) {numberOfSatellites = i;}
 	public void setIsSolid(boolean i) {isSolid = i;}
-	public void setMass(int i) {mass = i;}
+	public void setMass(double i) {mass = i;}
 	public void setAtmosphere(boolean i) {haveAtmosphere = i;}
 	public void setName(String i) {name = i;}
 	public void setPressure(double i) {pressure = i;}
@@ -35,7 +35,7 @@ public class Planeta {
 	public double getTemperature() {return temperature;}
 	public int getNumberOfSatellites() {return numberOfSatellites;}
 	public boolean getIsSolid() {return isSolid;}
-	public int getMass() {return mass;}
+	public double getMass() {return mass;}
 	public boolean getAtmosphere() {return haveAtmosphere;}
 	public String getName() {return name;}
 	public double getPressure() {return pressure;}
